fix(ci): unblock Tier-1 audit red — cppcheck FP + double-promotion
CI's Tier-1 audit has been red on every push since 2026-06-30: cppcheck
flagged a containerOutOfBounds on tests/test_contact_event.cpp:176
(events.front()), a false positive — the access is guarded by a fatal
ASSERT_FALSE(events.empty()) two lines up, which cppcheck cannot model
through the GoogleTest macro. Suppress it inline (--inline-suppr is on),
matching the documented precedent for this FP class (test_command_history.cpp,
see tools/audit/audit_config.yaml).

Also fix a real double-promotion clang-tidy surfaced in audio_loudness.cpp:40
(float gate constant implicitly widened to the libebur128 double accumulator)
— make the cast explicit.

+++ engine/audio/audio_loudness.cpp
@@ -37,7 +37,7 @@ float integratedLoudnessLufs(const int16_t* interleaved,
     static_assert(sizeof(int16_t) == sizeof(short),
                   "libebur128 short ingest assumes int16_t aliases short");
 
-    double lufs = kLoudnessSilenceGateLufs;
+    double lufs = static_cast<double>(kLoudnessSilenceGateLufs);
     if (ebur128_add_frames_short(
             st, reinterpret_cast<const short*>(interleaved), frames)
             == EBUR128_SUCCESS)
